+++ src/test/agrad/rev/matrix/softmax_test.cpp
@@ -49,7 +49,7 @@ softmax_grad(Eigen::Matrix<double,Eigen::Dynamic,1>& alpha_dbl,
     alpha(i) = alpha_dbl(i);
 
   std::vector<var> x(alpha.size());
-  for (int i = 0; i < x.size(); ++i)
+  for (size_t i = 0; i < x.size(); ++i)
     x[i] = alpha(i);
   
   var fx_k = stan::math::softmax(alpha)[k];
